Fix refresh rate config and allow donation block to be underground

+++ src/main/java/net/tropicraft/core/common/config/TropicsConfigs.java
@@ -63,7 +63,7 @@ public class TropicsConfigs {
     /* == Donations == */
     public static String tiltifyAppToken = "";
     public static int tiltifyCampaign = 0;
-    public static long donationTrackerRefreshRate = 10;
+    public static int donationTrackerRefreshRate = 10;
 
     private static final String C_GENERATION = "generation";
     private static final String C_ENTITY = "entity";
@@ -118,7 +118,7 @@ public class TropicsConfigs {
         
         tiltifyAppToken = config.get(C_DONATIONS, "tiltifyAppToken", tiltifyAppToken, "Add a token here to enable donation tracking, leave blank to disable").getString();
         tiltifyCampaign = config.get(C_DONATIONS, "tiltifyCampaign", tiltifyCampaign, "The tiltify campaign to track donations from").getInt();
-        donationTrackerRefreshRate = config.get(C_DONATIONS, "donationTrackerRefreshRate", donationTrackerRefreshRate, "How often the tracker checks for new donations, in seconds").getLong();
+        donationTrackerRefreshRate = config.get(C_DONATIONS, "donationTrackerRefreshRate", donationTrackerRefreshRate, "How often the tracker checks for new donations, in seconds").getInt();
         
         config.save();
     }
